\name{calm}
\alias{calm}
\docType{package}
\title{
Covariate Assisted Large-scale Multiple testing
}
\description{
Statistical methods for multiple testing with covariate information.
}
\details{
\tabular{ll}{
Package: \tab calm\cr
Type: \tab Package\cr
Version: \tab 0.9.0\cr
Date: \tab 2019-06-22\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Kun Liang <kun.liang@uwaterloo.ca>

Maintainer: Kun Liang <kun.liang@uwaterloo.ca>
}
\references{
Liang, K (2019) \emph{Empirical Bayes analysis of RNA sequencing experiments
with auxiliary information}.
}
\keyword{ package }
\seealso{
\code{\link{CLfdr}}
}
