% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calm.R
\name{calm.encode}
\alias{calm.encode}
\title{Encode Hypothesis Matrix into Coding Matrix}
\usage{
calm.encode(L, digits = 4)
}
\arguments{
\item{L}{A numeric hypothesis matrix (typically of size \code{g x h} where \code{g} is the number of groups).}

\item{digits}{Integer. Number of decimal places to round the result. Defaults to 4.}
}
\value{
A numeric matrix representing the coded design matrix. Row names are labeled \code{"GROUP #"}, and column names as \code{"CODE #"}.
}
\description{
Transforms a hypothesis matrix \code{L} into its corresponding coding matrix.
}
\details{
This function uses the Moore–Penrose inverse of \code{L} to compute the coding matrix.
It provides a contrast matrix that can be used for statistical model estimation.
}
