% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALM-package.R
\docType{package}
\name{CALM}
\alias{CALM}
\alias{CALM-package}
\title{CALM: Contrast Algorithms for Linear Models}
\description{
CALM provides a suite of functions to create and apply contrast codes for linear models.
}
\details{
CALM is an R package for creating and interpreting contrast codes for linear models.
It is an extension of the built-in contrast specifications in R and inspired by Venables \code{codingMatrices} package,
but it includes codes not available in other packages.
In addition to providing contrast specifications, it includes functions for directly converting
hypothesized differences (contrasts) into linear model codes (and vice versa).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwendorf/CALM}
  \item Report bugs at \url{https://github.com/cwendorf/CALM/issues}
}

}
\author{
\strong{Maintainer}: Craig Wendorf
\email{cwendorf@uwsp.edu}
}
