% File man/calmdecode.Rd
% Part of the CALM package

\name{calm.decode}
\alias{calm.decode}
\title{Decode Contrasts}
\description{Decode a set of contrast codes}

\usage{
\method{calm.decode}{default}(X,digits=4)
}

\arguments{
\item{X}{a matrix of contrast codes}
\item{digits}{number of decimal places for the codes}
}