% File man/calmencode.Rd
% Part of the CALM package

\name{calm.encode}
\alias{calm.encode}
\title{Encode Contrasts}
\description{Encode a set of linear contrasts}

\usage{
\method{calm.encode}{default}(L,digits=4)
}

\arguments{
\item{L}{a matrix of linear contrasts}
\item{digits}{number of decimal places for the codes}
}