% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{EstFDR}
\alias{EstFDR}
\title{FDR estimation}
\usage{
EstFDR(fdr)
}
\arguments{
\item{fdr}{vector of local FDR}
}
\value{
the estimate of the FDR
}
\description{
False discovery rate (FDR) estimation from local FDR
}
\examples{
lfdr <- c(runif(900), rbeta(100, 1, 10))
FDR <- EstFDR(lfdr)
sum(FDR<0.05)
}
