% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{EstNullProp_RB}
\alias{EstNullProp_RB}
\title{Right-boundary procedure}
\usage{
EstNullProp_RB(pval, lambda.vec = 0.05 * seq_len(19))
}
\arguments{
\item{pval}{vector of p-values}

\item{lambda.vec}{vector of lambda candidates (excluding 0 and 1)}
}
\value{
the estimate of the overall true null proportion
}
\description{
True null proportion (pi_0) estimator of Liang and Nettleton (2012), JRSSB
}
\examples{
pval <- c(runif(900), rbeta(100, 1, 10))
EstNullProp_RB(pval)
}
