% File man/CALM-package.Rd

\name{CALM-package}
\alias{CALM-package}
\docType{package}

\title{CALM}
\description{
\code{CALM} Contrast Algorithm for Linear Models
}

\details{
\tabular{ll}{
Package: \tab CALM\cr
Type: \tab Package\cr
}
Contrast Algorithms for Linear Models
}

\author{
Craig A. Wendorf\cr 
Department of Psychology\cr 
University of Wisconsin - Stevens Point\cr 
\email{cwendorf@uwsp.edu} 
}